refactor/wishlist-component
