+++ src/app/page.js
@@ -1,11 +1,12 @@
-import Footer from "@/Footer/Footer";
+
+import Footer from "@/components/footer/footer";
 import MainContent from "@/Main-Section/MainContent";
 
 export default function Home() {
   return (
     <div className="    ">
-      <MainContent/>
-      <Footer/>
+      <MainContent />
+      <Footer />
     </div>
   );
 }
